test(e2e): polish source related cases
