+++ e2e/cases/output/source-map/package.json
