+++ e2e/cases/output/source-map/rsbuild.config.ts
