+++ e2e/cases/output/source-map/src/App.jsx
