+++ e2e/cases/output/source-map/src/index.js
