--- e2e/cases/source/basic/src/common/test.js
@@ -1,1 +1,0 @@
-export const test = 1;
