--- e2e/cases/source/basic/src/index.js
@@ -1,9 +1,0 @@
-import { test } from '@common/test';
-
-const testEl = document.createElement('div');
-testEl.id = 'test';
-testEl.innerHTML = `Hello Rsbuild! ${test}`;
-
-document.body.appendChild(testEl);
-
-window.aa = 2;
