+++ e2e/cases/source/define/rsbuild.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from '@rsbuild/core';
+
+export default defineConfig({
+  source: {
+    define: {
+      ENABLE_TEST: JSON.stringify(true),
+    },
+  },
+});
