+++ e2e/cases/source/define/src/index.ts
