+++ e2e/cases/source/define/src/types.d.ts
