+++ e2e/cases/source/define/tsconfig.json
