+++ e2e/cases/source/pre-entry/rsbuild.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from '@rsbuild/core';
+
+export default defineConfig({
+  source: {
+    preEntry: ['./src/pre.js'],
+  },
+});
