+++ e2e/cases/source/pre-entry/src/index.js
@@ -1,0 +1,1 @@
+window.aa = 2;
