+++ e2e/cases/source/pre-entry/src/pre.js
